cmu-africa-assistant/requirements.txt

streamlit==1.29.0
openai==1.6.1
pinecone-client==3.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-pinecone==0.0.1
pandas==2.1.4
tiktoken==0.5.2
